streaming/pyproject.toml
[project]
name = "streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.17",
]
